moves finalresults gameinprogress joingame folders from components to pages
