+++ play/src/pages/gameinprogress/ChooseAnswer.tsx
@@ -6,8 +6,8 @@ import { GameSessionState } from '@righton/networking';
 import { Pagination } from 'swiper';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { BodyContentAreaDoubleColumnStyled } from '../../lib/styledcomponents/layout/BodyContentAreasStyled';
-import QuestionCard from '../QuestionCard';
-import AnswerCard from '../AnswerCard';
+import QuestionCard from '../../components/QuestionCard';
+import AnswerCard from '../../components/AnswerCard';
 import ScrollBoxStyled from '../../lib/styledcomponents/layout/ScrollBoxStyled';
 import 'swiper/css';
 import 'swiper/css/pagination';
