+++ play/src/pages/gameinprogress/DiscussAnswer.tsx
@@ -12,8 +12,8 @@ import { Swiper, SwiperSlide } from 'swiper/react';
 import { v4 as uuidv4 } from 'uuid';
 import { useTranslation } from 'react-i18next';
 import { AnswerState } from '../../lib/PlayModels';
-import QuestionCard from '../QuestionCard';
-import DiscussAnswerCard from '../DiscussAnswerCard';
+import QuestionCard from '../../components/QuestionCard';
+import DiscussAnswerCard from '../../components/DiscussAnswerCard';
 import ScrollBoxStyled from '../../lib/styledcomponents/layout/ScrollBoxStyled';
 import 'swiper/css';
 import 'swiper/css/pagination';
