+++ play/src/pages/joingame/HowToPlay.tsx
