+++ play/src/pages/joingame/SplashScreen.tsx
