+++ play/src/pages/joingame/howtoplayslides/HowToPlaySlide1Content.tsx
