+++ play/src/pages/joingame/howtoplayslides/HowToPlaySlide2Content.tsx
