+++ play/src/pages/joingame/howtoplayslides/HowToPlaySlide3Content.tsx
